fix(launchpad-g/erc20): move all zod schema to form.ts in utils folder
